+++ client/app/scripts/controllers/search/search-controller.js
@@ -1,0 +1,18 @@
+(function() {
+	'use strict';
+
+	/**
+	 * @ngdoc function
+	 * @name gemStoreApp.controller:SearchCtrl
+	 * @description
+	 * # SearchCtrl
+	 * Controller of the gemStoreApp
+	 */
+	angular.module('gemStoreApp')
+		.controller('SearchCtrl', ['$scope', function ($scope) {
+
+			$scope.clearSearch = function(){
+				$scope.search.name = '';
+			};
+	}]);
+})();