+++ client/app/scripts/directives/search/search-directive.js
@@ -11,8 +11,11 @@
 	 	.directive("search", function() {
 			return {
 				restrict: 'E',
-				templateUrl: 'scripts/directives/search/search.html'
+				templateUrl: 'scripts/directives/search/search.html',
+				controller: 'SearchCtrl'
 			};
+
+
 		});
 		
 })();