+++ client/app/scripts/services/products/products-service.js
@@ -1,18 +1,25 @@
 (function() {
     'use strict';
 
-	/**
-	 * @ngdoc service
-	 * @name gemStoreApp.productService
-	 * @description
-	 * # productService
-	 */
+    /**
+     * @ngdoc service
+     * @name gemStoreApp.productService
+     * @description
+     * # productService
+     */
 
-	angular.module('gemStoreApp.productService',['ngResource'])
+    //angular.module('gemStoreApp.productService',['ngResource','Restangular'])
+    angular.module('gemStoreApp.productService',['ngResource'])
 
-	    .factory('productsService', function($resource) {
-	        return $resource('/products/:id', {id:'@id'},{
-	        	'update': { method: 'PUT'}
-	        });
-	    });
+        //ngResource factory
+        .factory('productsService', function($resource) {
+            return $resource('/products/:id', {id:'@id'},{
+                'update': { method: 'PUT'}
+            });
+        });
+
+        //restAngular factory
+        // .factory('productsService', function(Restangular) {
+        //  return Restangular.service('products');
+        // });
 })();